+++ packages/devtools_app/lib/src/framework/home_screen.dart
@@ -295,7 +295,7 @@ class _ConnectInputState extends State<ConnectInput> with BlockingActionMixin {
         await FrameworkCore.initVmService(serviceUriAsString: uri);
     if (connected) {
       final connectedUri =
-          Uri.parse(serviceConnection.serviceManager.service!.wsUri!);
+          Uri.parse(serviceConnection.serviceManager.serviceUri!);
       routerDelegate.updateArgsIfChanged({'uri': '$connectedUri'});
       final shortUri = connectedUri.replace(path: '');
       notificationService.push('Successfully connected to $shortUri.');