+++ packages/devtools_app/lib/src/screens/memory/panes/diff/widgets/class_details/class_details.dart
@@ -57,7 +57,7 @@ class HeapClassDetails extends StatelessWidget {
       },
     );
 
-    return Split(
+    return SplitPane(
       axis: Axis.horizontal,
       initialFractions: const [0.7, 0.3],
       children: [