+++ packages/devtools_app/lib/src/screens/performance/panes/raster_stats/raster_stats.dart
@@ -131,7 +131,7 @@ class _LayerVisualizer extends StatelessWidget {
             ),
           );
         }
-        return Split(
+        return SplitPane(
           axis: Axis.horizontal,
           initialFractions: const [0.5, 0.5],
           children: [