+++ packages/devtools_app/lib/src/screens/vm_developer/isolate_statistics/isolate_statistics_view.dart
@@ -379,7 +379,7 @@ class _IsolatePortsWidgetState extends State<IsolatePortsWidget> {
   Widget build(BuildContext context) {
     final ports = widget.controller.ports;
     return OutlineDecoration(
-      child: Split(
+      child: SplitPane(
         axis: Axis.horizontal,
         initialFractions: const [
           0.3,