+++ packages/devtools_app/lib/src/screens/vm_developer/object_inspector/vm_instance_display.dart
@@ -79,7 +79,7 @@ class _VmInstanceDisplayState extends State<VmInstanceDisplay> {
 
   @override
   Widget build(BuildContext context) {
-    return Split(
+    return SplitPane(
       axis: Axis.vertical,
       initialFractions: const [0.5, 0.5],
       children: [