+++ packages/devtools_app/test/network/network_profiler_test.dart
@@ -104,13 +104,13 @@ void main() {
       expect(find.byType(ResumeButton), findsOneWidget);
       expect(find.byType(PauseButton), findsOneWidget);
       expect(find.byType(ClearButton), findsOneWidget);
-      expect(find.byType(Split), findsOneWidget);
+      expect(find.byType(SplitPane), findsOneWidget);
 
       // Advance the clock to populate the network requests table.
       await tester.pump(const Duration(seconds: 2));
       expect(find.byType(CircularProgressIndicator), findsNothing);
 
-      expect(controller.requests.value.requests, isNotEmpty);
+      expect(controller.requests.value, isNotEmpty);
     }
 
     // We should see the list of requests and the inspector, but have no
@@ -275,7 +275,7 @@ void main() {
           }
         }
 
-        for (final request in controller.requests.value.requests) {
+        for (final request in controller.requests.value) {
           controller.selectedRequest.value = request;
           await tester.pumpAndSettle();
           expect(find.text('No request selected'), findsNothing);