+++ tool/lib/commands/build.dart
@@ -95,6 +95,8 @@ class BuildCommand extends Command {
             '--web-renderer',
             'canvaskit',
             '--pwa-strategy=offline-first',
+            // Enable default optimizations: https://dart.dev/tools/dart-compile#js
+            '--dart2js-optimization=O1',
             if (buildMode != 'debug') '--$buildMode',
             '--no-tree-shake-icons',
           ],